Add a fallback for players moving surfaces
Space Exploration has a navigation mode that creates a fake
player on other surfaces to view them from a 'satellite'. In the
existing code, any time that there isn't a character body, the robots
don't follow. Unfortunately, there also isn't a body for some cutscenes
that happen on transitions to other worlds.

Use the sanctioned interface to check if we're using the sat view vs
actually arriving somewhere else in a cutscene.

There are more sat view bugs, but not leaving the bots behind makes
them less likely to occur.

Credit to @jmerdich